+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Allow to use extended capabilities
+* Allow to send and receive custom proto packet
+
 ## 2022-05-01 v1.1.0
 
 * Allow to interrupt character manually